final touchups
