+++ src/components/SearchResultsTable.vue
@@ -118,7 +118,6 @@ export default {
                 viewButton.innerHTML = 'View';
                 viewButton.onclick = function() {
                     console.log('this is the partyquest id', docs.id);
-                    alert('pause the load');
                     window.location.replace('/pq/' + docs.id);
                 };
                 // add button to action column cell
