+++ src/components/pq_details_page/PqDetails.vue
