--- src/views/ChatsPage.vue
@@ -1,88 +1,0 @@
-<template>
-    <chat-container />
-</template>
-
-<script>
-import ChatContainer from '../components/chats_page/ChatContainer.vue';
-export default {
-    name: 'ChatsPage',
-    components: {
-        ChatContainer,
-    },
-};
-</script>
-<style>
-body {
-    background: url('../assets/pubnub-filter-blocks-bg.png');
-}
-
-#app {
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-    text-align: center;
-    color: #454d54;
-    margin-top: 60px;
-}
-
-/* Loading Ripple Animation */
-/* Generated by loading.io */
-@keyframes lds-ripple {
-    0% {
-        top: 96px;
-        left: 96px;
-        width: 0;
-        height: 0;
-        opacity: 1;
-    }
-    100% {
-        top: 8px;
-        left: 8px;
-        width: 176px;
-        height: 176px;
-        opacity: 0;
-    }
-}
-@-webkit-keyframes lds-ripple {
-    0% {
-        top: 96px;
-        left: 96px;
-        width: 0;
-        height: 0;
-        opacity: 1;
-    }
-    100% {
-        top: 8px;
-        left: 8px;
-        width: 176px;
-        height: 176px;
-        opacity: 0;
-    }
-}
-.lds-ripple {
-    position: relative;
-}
-.lds-ripple div {
-    box-sizing: content-box;
-    position: absolute;
-    border-width: 4px;
-    border-style: solid;
-    opacity: 1;
-    border-radius: 50%;
-    -webkit-animation: lds-ripple 2.6s cubic-bezier(0, 0.2, 0.8, 1) infinite;
-    animation: lds-ripple 2.6s cubic-bezier(0, 0.2, 0.8, 1) infinite;
-}
-.lds-ripple div:nth-child(1) {
-    border-color: #000000;
-}
-.lds-ripple div:nth-child(2) {
-    border-color: rgba(0%, 0%, 0%, 0.675);
-    -webkit-animation-delay: -1.3s;
-    animation-delay: -1.3s;
-}
-.lds-ripple {
-    width: 36px !important;
-    height: 36px !important;
-    -webkit-transform: translate(-18px, -18px) scale(0.18) translate(18px, 18px);
-    transform: translate(-18px, -18px) scale(0.18) translate(18px, 18px);
-}
-</style>
