+++ src/views/EditProfilePage.vue
@@ -362,7 +362,6 @@ export default {
                 });
                 console.log(docRef);
                 document.getElementById('editprofileform').reset();
-                alert('Personal Information Updated!');
                 this.goBack();
             } catch (error) {
                 console.error('Error updating profile: ', error);
